Changed dtos/model classes and dates are now LocalDateTime
