+++ Project/app/src/main/java/org/ionproject/android/class_section/JournalsAdapter.kt
@@ -6,7 +6,7 @@ import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import kotlinx.android.synthetic.main.list_item_label.view.*
 import org.ionproject.android.R
-import org.ionproject.android.common.dtos.JournalSummary
+import org.ionproject.android.common.model.JournalSummary
 
 class JournalsAdapter(
     private val model: ClassSectionViewModel
