+++ Project/app/src/main/java/org/ionproject/android/class_section/WorkAssignmentsAdapter.kt
@@ -6,7 +6,7 @@ import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import kotlinx.android.synthetic.main.list_item_label.view.*
 import org.ionproject.android.R
-import org.ionproject.android.common.dtos.TodoSummary
+import org.ionproject.android.common.model.TodoSummary
 
 class WorkAssignmentsAdapter(
     private val model: ClassSectionViewModel
