+++ Project/app/src/main/java/org/ionproject/android/common/model/Journal.kt
@@ -1,12 +1,13 @@
-package org.ionproject.android.common.dtos
+package org.ionproject.android.common.model
 
-import org.ionproject.android.common.siren.Journal
+import org.ionproject.android.common.dto.Journal
+import java.time.LocalDateTime
 
 class JournalSummary(
     val uid: String,
     val summary: String?,
     val description: String?,
-    val lastModification: String?
+    val lastModification: LocalDateTime?
 )
 
 fun Journal.toJournalSummary() =
@@ -14,5 +15,5 @@ fun Journal.toJournalSummary() =
         uid = properties.uid.value[0],
         summary = properties.summary?.value?.get(0),
         description = properties.description?.value?.get(0),
-        lastModification = properties.dtstamp?.value?.get(0)
+        lastModification = properties.dtstamp?.value?.get(0)?.toLocalDate()
     )
